Fix DC2_exp.ipynb and delete some redundant files (#985) * fix DC2_exp.ipynb, and it can now reproduce the results
* move dc2_new.py to dc2.py

* write a new funtion get_test_sample for DC2 to generate sample for plotting

* delete config.yaml; it's an old config

* delete DC2_deconv.ipynb

* move DC2_exp_fixed.ipynb to DC2_exp.ipynb

* delete ori_config.yaml, test_config.yaml and train_config.yaml; they are old config

* add notebook_config.yaml for DC2_exp.ipynb

* edit train_dc2.sh

---------
