--- case_studies/dc2/test_config.yaml
@@ -1,36 +1,0 @@
----
-defaults:
-    - ../../bliss/conf@_here_: base_config
-    - _self_
-    - override hydra/job_logging: stdout
-
-encoder:
-    survey_bands: ["g", "i", "r", "u", "y", "z"]
-    tile_slen: 4
-    double_detect: false
-    tiles_to_crop: 0
-    do_data_augmentation: true # set this according to your training model
-    image_normalizer:
-        bands: [0, 1, 2, 3, 4, 5]
-        concat_psf_params: true # set this according to your training model
-        num_psf_params: 4  # for SDSS, 4 for DC2
-        log_transform_stdevs: []
-        use_clahe: false
-
-
-surveys:
-    dc2:
-        data_dir: ${paths.data}/tests/dc2/dc2_multiband/
-        cat_path: ${paths.dc2}/merged_catalog/merged_catalog_psf_100000.pkl
-        n_split: 1
-        image_lim: [4000, 4000]
-        for_plotting: true
-
-train:
-    trainer:
-        logger:
-            name: DC2
-            version: DC2_origin
-        precision: 32
-    data_source: ${surveys.dc2}
-    pretrained_weights: null
